First push of the code.
The game is playable, but there are some things that could do with
fixing. Tests need to be added. If there are too many pieces on a
position, they should stack in layers. At the end of the game, the
board should reset to let you have another game. etc.

The code could also do with a clean-up, commenting, etc.